Introduce IComponentDisposer to manually call dispose
