+++ Source/Blazorise/ComponentActivator.cs
@@ -14,14 +14,6 @@ namespace Blazorise
     /// </summary>
     public class ComponentActivator : IComponentActivator
     {
-        #region Members
-
-        private bool disposePossible;
-
-        private readonly IList<object> disposables;
-
-        #endregion
-
         #region Constructors
 
         /// <summary>
@@ -31,7 +23,6 @@ namespace Blazorise
         public ComponentActivator( IServiceProvider serviceProvider )
         {
             ServiceProvider = serviceProvider;
-            disposables = LoadServiceProviderDisposableList();
         }
 
         #endregion
@@ -45,9 +36,6 @@ namespace Blazorise
         /// <returns>Return the newly created component or raises an exception if the specified typo is invalid.</returns>
         public IComponent CreateInstance( Type componentType )
         {
-            if ( disposePossible )
-                RemoveDisposedComponents();
-
             var instance = ServiceProvider.GetService( componentType );
 
             if ( instance == null )
@@ -61,36 +49,6 @@ namespace Blazorise
             }
 
             return component;
-        }
-
-        /// <summary>
-        /// Loads the ServiceProvider's list of disposables
-        /// </summary>
-        /// <returns>List of object references the ServiceProvider uses to track disposables</returns>
-        private IList<object> LoadServiceProviderDisposableList()
-        {
-            var disposablesPropertyInfo = ServiceProvider.GetType().GetProperty( "Disposables", BindingFlags.Instance | BindingFlags.NonPublic );
-
-            disposePossible = disposablesPropertyInfo is not null;
-
-            return disposablesPropertyInfo?.GetValue( ServiceProvider ) as IList<object> ?? Array.Empty<object>();
-        }
-
-        /// <summary>
-        /// Removes all disposed Blazorise components from the ServiceProvider's disposables list.
-        /// This prevents a memory leak where these references were being held permanently.
-        /// </summary>
-        private void RemoveDisposedComponents()
-        {
-            var disposedComponents = disposables
-                .OfType<BaseAfterRenderComponent>()
-                .Where( x => x.Disposed || x.AsyncDisposed )
-                .ToList();
-
-            foreach ( var component in disposedComponents )
-            {
-                disposables.Remove( component );
-            }
         }
 
         #endregion
